% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{AlphaFoldDetails}
\alias{AlphaFoldDetails}
\title{AlphaFoldDetails shiny element}
\usage{
AlphaFoldDetails(class = NULL, afId = NULL, showFiles = FALSE, showPae = FALSE)
}
\arguments{
\item{afId}{a character string containing a UniProt id that is in the
AlphaFold database.}

\item{showFiles}{a logical value indicating if the download links to the
files in the database should be shown in a div.}

\item{showPae}{a logical value indicating if the predicted aligned
error image should be shown.}
}
\description{
AlphaFoldDetails shiny element
}
\examples{
AlphaFoldDetails(afId = "A0A1U8FD60", showPae = TRUE, showFiles = TRUE)
}
