% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{AlphaFoldMolstar}
\alias{AlphaFoldMolstar}
\title{AlphaFoldMolstar shiny element}
\usage{
AlphaFoldMolstar(...)
}
\arguments{
\item{...}{Component props and children. See the official Blueprint docs for
details.}
}
\description{
AlphaFoldMolstar shiny element
}
\examples{
url <- "https://alphafold.ebi.ac.uk/files/AF-A0A1U8FD60-F1-model_v4.pdb"
AlphaFoldMolstar(afId = "A0A1U8FD60", dimensions=c(300, 300))
}
