% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_utils.R
\name{footer_tag}
\alias{footer_tag}
\title{Common footer for all examples}
\usage{
footer_tag()
}
\value{
a taglist for the footer and css
}
\description{
Common footer for all examples
}
\examples{
footer_tag()
}
