% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_utils.R
\name{footerTag}
\alias{footerTag}
\title{Common footer for all examples}
\usage{
footerTag()
}
\value{
a common footer and css for the examples
}
\description{
This function is not exported and shouldn't be used outside of the examples
}
