% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_utils.R
\name{get_sample_code}
\alias{get_sample_code}
\title{Get minimal sample code from example file}
\usage{
get_sample_code(lookup, example_path)
}
\arguments{
\item{lookup}{suffix of tag that surrounds code}

\item{example_path}{path to file to load example from}
}
\value{
a vector of lines of the
}
\description{
It removes lines that are surrounded by a START and END tag.
}
\details{
There cannot be ambiguous tags in the code or the same tag appearing
multiple times.

For example, the vector with the following lines:

\if{html}{\out{<div class="sourceCode">}}\preformatted{... # code before
# START block1
... # some code
# END block1
... # code after
}\if{html}{\out{</div>}}

will be changed to

\if{html}{\out{<div class="sourceCode">}}\preformatted{... # code before
... # code after
}\if{html}{\out{</div>}}
}
\examples{
get_sample_code("remove_from_sample_ui", "examples/AlphaFoldDetails.R")

# Removes two blocks of code
get_sample_code(
  c(
    "remove_from_sample_ui",
    "remove_from_sample_server"
  ),
  "examples/AlphaFoldDetails.R"
)
}
