% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_utils.R
\name{getSampleCode}
\alias{getSampleCode}
\title{Get minimal sample code from example file}
\usage{
getSampleCode(lookup, examplePath)
}
\arguments{
\item{lookup}{suffix of tag that surrounds code}

\item{examplePath}{path to file to load example from}
}
\value{
a vector of lines of the sample code
}
\description{
It removes lines that are surrounded by a "# START \if{html}{\out{<tag>}}" and "# END \if{html}{\out{<tag>}}"
comment.
}
\details{
There cannot be ambiguous tags in the code or the same tag appearing
multiple times.

For example, the vector with the following lines:

\if{html}{\out{<div class="sourceCode">}}\preformatted{... # code before
# START block1
... # some code
# END block1
... # code after
}\if{html}{\out{</div>}}

will be changed to

\if{html}{\out{<div class="sourceCode">}}\preformatted{... # code before
... # code after
}\if{html}{\out{</div>}}
}
