% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{template}
\alias{template}
\title{template}
\arguments{
\item{inputId}{The `input` slot that will be used to access the value.}

\item{...}{Component props and children. See the official Blueprint docs for
details.}

\item{value}{Initial value.}
}
\description{
template
}
