% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{AlphaFoldDetails}
\alias{AlphaFoldDetails}
\title{AlphaFoldDetails shiny element}
\usage{
AlphaFoldDetails(...)
}
\arguments{
\item{...}{Component props and children. See the official Blueprint docs for
details.}
}
\description{
AlphaFoldDetails shiny element
}
\examples{
url <- "https://alphafold.ebi.ac.uk/files/AF-A0A1U8FD60-F1-model_v4.pdb"
AlphaFoldMolstar(url = url, dimensions=c(300, 300))
}
