% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_utils.R
\name{footer_tag}
\alias{footer_tag}
\title{Common footer for all examples}
\usage{
footer_tag()
}
\value{
a common footer and css for the examples
}
\description{
This function is not exported and shouldn't be used outside of the examples
}
