% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Molstar}
\alias{Molstar}
\title{Molstar shiny element}
\usage{
Molstar(...)
}
\arguments{
\item{...}{Component props and children. See the official Blueprint docs for
details.}
}
\description{
Molstar shiny element
}
\examples{
Molstar(pbdId="1LOL", dimensions=c(300, 300))
}
